Fix cppcheck error

+++ tests/unit/s2n_mem_usage_test.c
@@ -237,8 +237,8 @@ int main(int argc, char **argv)
     if (allocation_diff > maxAllowedMemDiff
             || handshake_diff > maxAllowedMemDiff
             || handshake_diff < minAllowedMemDiff) {
-        fprintf(stdout, "\nActual KB per connection: %zd\n",
-                ACTUAL_MEM_PER_CONNECTION(connectionsToUse, handshake_diff));
+        fprintf(stdout, "\nActual KB per connection: %i\n",
+                (int) ACTUAL_MEM_PER_CONNECTION(connectionsToUse, handshake_diff));
         FAIL_MSG("Unexpected memory usage. If expected, update MEM_PER_CONNECTION.");
     }
